+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##  [Unreleased]
 ###  Added
+- Remove all items from shopping list.
+- Count items from shopping list.
+- List items from shopping list.
 - Use logrus as logging system.
 - Remove an item from the shopping list (with reprompt).
 - Add an item to the shopping list (with reprompt).